java: bcv-cfr

+++ app/src/main/java/javax/microedition/media/control/ToneControl.java
@@ -1,17 +1,23 @@
+/*
+ * Decompiled with CFR.
+ * 
+ * Could not load the following classes:
+ *  javax.microedition.media.Control
+ */
 package javax.microedition.media.control;
 
 import javax.microedition.media.Control;
 
-/* loaded from: classes.dex */
-public class ToneControl implements Control {
-    public static byte BLOCK_END = -6;
-    public static byte BLOCK_START = -5;
-    public static byte C4 = 60;
-    public static byte PLAY_BLOCK = -7;
-    public static byte REPEAT = -9;
-    public static byte RESOLUTION = -4;
-    public static byte SET_VOLUME = -8;
-    public static byte SILENCE = -1;
-    public static byte TEMPO = -3;
-    public static byte VERSION = -2;
+public class ToneControl
+implements Control {
+    public static byte BLOCK_END = (byte)-6;
+    public static byte BLOCK_START = (byte)-5;
+    public static byte C4 = (byte)60;
+    public static byte PLAY_BLOCK = (byte)-7;
+    public static byte REPEAT = (byte)-9;
+    public static byte RESOLUTION = (byte)-4;
+    public static byte SET_VOLUME = (byte)-8;
+    public static byte SILENCE = (byte)-1;
+    public static byte TEMPO = (byte)-3;
+    public static byte VERSION = (byte)-2;
 }
